Order listening services better
With the recent additions of multiple-listen-service ordering, the order
using the concat-fragment depended on the end-of-line vs other
alphanumeric characters. This is not actually the case with the concat
module, so this fixes ordering for services under a small set of edge
cases.

This adds the smaller edge case of if a balancermember begins with "00".