Allow the fact to manage but disable the haproxy service
In an active/passive scenario, as a user I'd like the
configuration to be set and the package to be installed
without the service being started.

The `if $enable` (l81) prevents this. Today, if one sets
`manage_service = true` and `enable = false`, it will result
in a catalog compilation issue, due to the `Concat[/etc/haproxy/haproxy.cfg']`
dependency on line 134 not being met.