Avoid mixing up backend servers
If an haproxy::listen and haproxy::backend have the same name, their
order will be `20-<name>-00` and concat ordering will be mixed up. The
haproxy::balancermember resource uses this ordering to add backend
servers to a listen/backend block (with the order `20-<name>-01`) and
this is the least-impactful way to avoid balancer members from being
cross-associated without introducing backwards-incompatible changes.