Implement `options` as array of hashes so order is preserved
Fixes https://tickets.puppetlabs.com/browse/MODULES-1317. See also the
discussion in PR #171.

Frontend, backend and listen options are specified via the
`haproxy::frontend::options`, `haproxy:backend::options` and
`haproxy::listen::options` parameters. Until now this `options`
parameter was a hash, which means it had to be sorted before processing
it in the `_options.erb` template fragment (since Ruby 1.8.x has
non-deterministic hash key ordering).

This clashes with the fact that in HAProxy's configuration order matters.

With this change we allow the `options` parameter to be either a hash,
like before, or an array of hashes, where each hash represents a
frontend, backend or listen option and its value. This way the order in
which options will appear in the final configuration file can be
controlled entirely by the user.

This change is completely backwards compatible and contains updated spec
tests and documentation.