Fix determining $haproxy::config_dir in haproxy::instance
The haproxy::mapfile define requires that File[$haproxy::config_dir] is
managed (because that's where the map files live) but the way a
single-instance HAProxy deployment is created ignores the class
parameter $haproxy::config_dir (which is set according to platform by
default). This can result in a failing Puppet run because the config dir
doesn't (yet) exist when the map files are written.

This also means for osfamily == Gentoo we no longer need to explicitly
create /etc/haproxy (= $haproxy::config_dir).