Re-add support for specifying package version in package_ensure
Restricting haproxy::package_ensure to one of 'present', 'absent' or
'latest' prevents the user from letting Puppet install a specific
version of the HAProxy package. This restriction is unwise. It is also
an unnecessary breaking change from version 1.5.0.

This commit replaces the Enum data type for haproxy::package_ensure
and haproxy::instance::package_ensure with the String data type, adds
spec tests for setting a specific package version and updates the
documentation.