Override  $sort_options_alphabetic when 'httpchk' option defined
When using the option 'httpchk', it must be positioned before the
'http-check' directive in haproxy.cfg otherwise it will be ignored.

A warning will be output to stderr to inform the user that the
value of $sort_options_alphabetic is being overridden.

Also added an acceptance to ensure the expected ordering is taking
place and a warning is presented.

Add is_hash check for $options param in haproxy::backend