backend: dont log warnings if not necessary
this is a follow-up for https://github.com/puppetlabs/puppetlabs-haproxy/pull/442

with the current code in the main branch, a warning is logged to the
puppetserver.log for every backend, if httpchk is set. This is very
noisy and provides 0 benefit, if the user set the
sort_options_alphabetic to false.