Fix 'option' entry name in option_order hash
The `option_order` hash is used to determine the correct order of
various backend/listen configuration parameters but it contained
a misspelled "options" parameter that is really called "option". This
change fixes that and extends the spec tests for the haproxy::listen
define to catch this problem.

See the discussion in #442 for more details and the reasoning behind the
`option_order` hash and why option order is important in HAProxy.
Unfortunately the change in #442 also introduced the misspelled
parameter and no spec test caught it.